requirements.txt
alpaca-py==0.39.1
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
curl_cffi==0.10.0
frozendict==2.4.6
idna==3.10
msgpack==1.1.0
multitasking==0.0.11
numpy==2.2.4
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
sseclient-py==1.8.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
websockets==15.0.1
yfinance==0.2.58
